backend/requirements.txt
# Core dependencies
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
requests>=2.31.0
wikipedia>=1.4.0
google-generativeai>=0.3.0

# Database dependencies
sqlalchemy>=2.0.0
alembic>=1.12.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Utilities
python-multipart>=0.0.6
python-jose>=3.3.0
passlib>=1.7.4
email-validator>=2.1.0
